adjust test location
